requirements.txt
flask
randomwords
requests
sqlitedict
protobuf
python-snappy